requirements.txt
asgiref==3.7.2
AshLogger==0.3
click==8.1.5
colorama==0.4.6
Django==4.2.3
django-auditlog==2.3.0
django-cors-headers==4.2.0
django-ipware==5.0.0
django-user-agents==0.4.0
django4-background-tasks==1.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
h11==0.14.0
httpagentparser==1.9.5
# mysqlclient==2.2.0
pika==1.3.2
PyJWT==2.7.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
ua-parser==0.18.0
user-agents==2.2.0
uvicorn==0.23.0
